Restructured router setup with different files and installed argon2
