+++ package.json
@@ -13,6 +13,7 @@
   "dependencies": {
     "@types/config": "^3.3.0",
     "@types/jsonwebtoken": "^9.0.1",
+    "argon2": "^0.30.3",
     "config": "^3.3.9",
     "connect-redis": "^7.0.1",
     "cors": "^2.8.5",
