+++ server/routers/baseRouter.ts
@@ -1,0 +1,5 @@
+import express from 'express';
+
+const router = express.Router();
+
+module.exports = router;
